# Fix for SurvIT_ode() in case mcmc_size=1