… NAMESPACE
… src/.gitignore