Currently, SurvIT_ode() fails if mcmc_site==1 because the apply-function expects a matrix. Thanks.